A tentative fix for brightness explossion problem

+++ include/slg/samplers/metropolis.h
@@ -46,7 +46,7 @@ public:
 	// I'm storing totalLuminance and sampleCount on shared variables
 	// in order to have far more accurate estimation in the image mean intensity
 	// computation
-	double totalLuminance, sampleCount;
+	float totalLuminance, sampleCount;
 };
 
 //------------------------------------------------------------------------------
